switch init to lua, use lazy.nvim and more
I switched init to lua since that's the future and based on a quick
look, it appears there are more source on Lua than Vim script.

VimPlug got replaced in favor of the more modern lazy.nvim. One visible
difference is that VimPlug has to be installed on the host before nvim
can be used, whereas the latter doesn't.

I also stopped using gruvbox and switched to One Dark since it's easier
on the eyes.

I am very happy to start using NvimTree as well, making the file
navigation much easier! One thing to add soon is shortcuts to open and
close it.

Lastly I started using vim-sleuth so that indentation and other similar
rules are set based on the rest of the project automatically.