change: use different auto-save, keymapping improvements
The previous auto-save worked fine 95% of the time, but the rest 5% were
pissing me off. Basically auto-save would save in situation where it
wasn't desired, when I was in Insert mode usually. I tried to tweak its
config but for some reason my changes were never accepted, though I
followed their docs. In the end, I gave up on it and switched to its
fork which is more actively maintained.