change: switch vscode nvim init to Lua
I had quite a lot of trouble with <C-L> and other split nav commands.
VSC just wouldn't accept them for the love of god. I couldn't find
anything out of ordinary with the mappings in the init file nor in the
keybindings on the VSC's side. In the end, I mapped them not to <C-W>L
but rather to the explicit VSC commands which solved the issue.